Extract schema output methods to schema_output.go
